Address more comments, refactor testutil folder to util, move constant from Utility class.
