+++ src/main/java/com/rocketden/main/util/Utility.java
@@ -2,12 +2,9 @@ package com.rocketden.main.util;
 
 import java.util.Random;
 
-import com.rocketden.main.controller.v1.BaseRestController;
-
 public class Utility {
 
     private static final Random random = new Random();
-    public static final String SOCKET_PATH = BaseRestController.BASE_SOCKET_URL + "/%s/subscribe-user";
 
     // Generate numeric String with a specific length.
     public String generateId(int length) {
